+++ Thunderstore/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2026-05-07
+
+### Fixed
+
+- Quota rollover resetting to 0
+- Deadline days now sync correctly to all clients after quota fulfillment
+- _GeneralImprovements_ - Total Days and Total Quotas monitors now show correct values after rejoining
+- _Advanced Features_ - Crew wipe now shows the real scrap loss instead of always _Lost 100% scrap_
+
 ## [1.0.1] - 2026-05-06
 
 ### Fixed